+++ armadillo/src/test/java/org/molgenis/armadillo/service/ManagementServiceTest.java
@@ -17,8 +17,6 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.CopyOnWriteArrayList;
-import java.util.concurrent.TimeUnit;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -359,7 +357,7 @@ class ManagementServiceTest {
     m.setAccessible(true);
 
     String jarHome = (String) m.invoke(service);
-    assertTrue(jarHome.endsWith("build/libs/"));
+    assertTrue(jarHome.endsWith("build/libs"));
   }
 
   @Test
@@ -631,77 +629,6 @@ class ManagementServiceTest {
   }
 
   @Test
-  void thread_should_have_correct_name_and_be_daemon() throws Exception {
-    Thread tailer = service.startLogTailer(logFile, line -> {});
-
-    assertThat(tailer.getName()).isEqualTo("update-log-tailer");
-    assertThat(tailer.isDaemon()).isTrue();
-    assertThat(tailer.isAlive()).isTrue();
-
-    tailer.interrupt();
-  }
-
-  @Test
-  void should_pick_up_lines_written_after_start() throws Exception {
-    List<String> captured = new CopyOnWriteArrayList<>();
-
-    Thread tailer = service.startLogTailer(logFile, captured::add);
-
-    // Write lines AFTER the tailer has started (it skips existing content)
-    await().atMost(500, TimeUnit.MILLISECONDS).until(() -> tailer.isAlive());
-
-    try (PrintWriter writer = new PrintWriter(new FileWriter(logFile, true))) {
-      writer.println("line one");
-      writer.println("line two");
-      writer.flush();
-    }
-
-    await().atMost(2, TimeUnit.SECONDS).until(() -> captured.size() >= 2);
-
-    assertThat(captured).containsExactly("line one", "line two");
-
-    tailer.interrupt();
-  }
-
-  @Test
-  void should_skip_content_already_in_file_at_start() throws Exception {
-    // Write content BEFORE starting the tailer
-    Files.writeString(logFile.toPath(), "pre-existing line\n");
-
-    List<String> captured = new CopyOnWriteArrayList<>();
-    Thread tailer = service.startLogTailer(logFile, captured::add);
-
-    // Give it time to potentially (wrongly) pick up the old line
-    Thread.sleep(300);
-
-    assertThat(captured).isEmpty();
-
-    tailer.interrupt();
-  }
-
-  @Test
-  void should_stop_cleanly_on_interrupt() throws Exception {
-    Thread tailer = service.startLogTailer(logFile, line -> {});
-
-    tailer.interrupt();
-
-    await().atMost(1, TimeUnit.SECONDS).until(() -> !tailer.isAlive());
-
-    assertThat(tailer.isAlive()).isFalse();
-  }
-
-  @Test
-  void should_log_error_on_missing_file() throws Exception {
-    File missing = tempDir.resolve("nonexistent.log").toFile();
-
-    Thread tailer = service.startLogTailer(missing, line -> {});
-
-    tailer.join(5000); // wait up to 5s for the thread to die naturally
-
-    assertThat(tailer.isAlive()).isFalse();
-  }
-
-  @Test
   void getPercentage_should_calculate_correctly() {
     assertThat(service.getPercentage(50, 100)).isEqualTo(50);
     assertThat(service.getPercentage(1, 100)).isEqualTo(1);
@@ -740,14 +667,6 @@ class ManagementServiceTest {
     when(lastReleaseResponse.statusCode()).thenReturn(404);
 
     assertThatThrownBy(() -> service.getLastRelease()).isInstanceOf(ResponseStatusException.class);
-  }
-
-  @Test
-  void getProcessBuilderForRebootScript() throws Exception {
-    String pythonScript = "print('hello world')";
-    ProcessBuilder pb = service.getProcessBuilderForRebootScript(pythonScript);
-    assertThat(pb.command().get(2)).isEqualTo(pythonScript);
-    assertThat(pb.redirectInput().file()).isEqualTo(new File("/dev/null"));
   }
 
   @Test
